# api-spec/openapi.json
{"openapi":"3.1.0","info":{"title":"BoostWrapper API","version":"0.1.0"},"paths":{"/train":{"post":{"summary":"Train","operationId":"train_train_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainInput"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/info/{task_id}":{"get":{"summary":"Get Info","operationId":"get_info_info__task_id__get","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/predict":{"post":{"summary":"Predict","operationId":"predict_predict_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PredictByIdInput"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PredictByIdInput":{"properties":{"task_id":{"type":"string","title":"Task Id"},"X":{"items":{"items":{"type":"number"},"type":"array"},"type":"array","title":"X"}},"type":"object","required":["task_id","X"],"title":"PredictByIdInput"},"TrainInput":{"properties":{"X":{"items":{"items":{"type":"number"},"type":"array"},"type":"array","title":"X"},"y":{"items":{"type":"number"},"type":"array","title":"Y"},"backend":{"type":"string","enum":["catboost","xgboost","lightgbm"],"title":"Backend"},"task":{"type":"string","enum":["classification","regression"],"title":"Task"},"params":{"additionalProperties":true,"type":"object","title":"Params"}},"type":"object","required":["X","y","backend","task"],"title":"TrainInput"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}